// src/components/sux-button/interfaces.ts
export type ButtonColor = "blue" | "inverse" | "neutral" | "red";
export type ButtonVariant = 'cta' | 'primary' | 'secondary' | 'negative';
export type ButtonSize = 's' | 'm' | 'l' | 'xl';
export type ButtonAlignment =
  | "start"
  | "end"
  | "center"
  | "space-between"
  | "icon-start-space-between"
  | "icon-end-space-between";

export type DropdownIconType = "chevron" | "caret" | "ellipsis" | "overflow";